requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
protobuf==3.20.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.6
six==1.16.0
typing_extensions==4.4.0
tzdata==2022.7
Werkzeug==2.2.2
zipp==3.10.0
